requirements.txt
streamlit
google-generativeai
SpeechRecognition
pyttsx3
geopy
opencv-python
numpy
Pillow
